requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.0.7
django-cors-headers==3.4.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
html5lib==1.1
idna==2.10
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
webencodings==0.5.1
